% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_HS.R
\name{eval_HS}
\alias{eval_HS}
\title{Slice sampler for the Normal observations model with HS-like prior.
 TODO: More detail on priors.}
\usage{
eval_HS(Y)
}
\description{
@param Y observations
 @return A list of samples
 @export
}
