% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logshrink.R
\name{eval_hslike}
\alias{eval_hslike}
\title{Slice sampler for the Normal observations model with HS-like prior.
 TODO: More detail on priors.}
\usage{
eval_hslike(Y, nmc = 100, burn = 100, eff_zero = 1e-05, verbose = TRUE)
}
\description{
@param Y observations
 @return A list of samples
 @export
}
