% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hslike.r
\docType{package}
\name{hslike}
\alias{hslike}
\alias{hslike-package}
\title{hslike: A package for estimation with the Horseshoe-like prior.}
\description{
hslike: A package for estimation with the Horseshoe-like prior.
}
\references{
\emph{The Horseshoe+ Estimator of Ultra-Sparse Signals}
Anindya Bhadra, Jyotishka Datta, Nicholas G. Polson, Brandon Willard
\cr
\url{http://arxiv.org/abs/1502.00560}
\url{https://github.com/DattaHub/hslike}
}
